+++ part2/reverse_flask_reza/project/templates/403.html
@@ -1,0 +1,13 @@
+{% extends "index.html" %}
+
+{% block content %}
+
+<div class="error">
+
+  <h1>Error 403</h1>
+  <p>Access denied.</p>
+  <p><em>Return <a href="{{url_for('home')}}">Home</a></em></p>
+  
+</div>
+
+{% end block %}
