+++ part2/reverse_flask_reza/project/templates/500.html
@@ -1,0 +1,14 @@
+{% extends "index.html" %}
+
+{% block content %}
+
+<div class="error">
+
+    <h1>500</h1>
+    <p>Something is wrong! We are working on it!</p>
+    <p><em>Return <a href="{{url_for('home')}}">Home.</a></em></p>
+
+</div>
+
+
+{% endblock %}
