+++ part2/reverse_flask_reza/project/templates/index.html
@@ -2,19 +2,19 @@
 
 <html>
   <head>
-    <title>Flask - Behavior Driven Development</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="http://netdna.bootstrapcdn.com/bootstrap/3.1.1/css/bootstrap.min.css" rel="stylesheet" media="screen">
+    <title>Reversing string.</title>
+	<link rel="stylesheet" type="text/css" href="main.css">
+	
   </head>
   <body>
     <div class="container">
-      <!-- inheritance -->
-      {% block content %}
-      {% endblock %}
-      <!-- end inheritance -->
-    </div>
-    <script src="http://code.jquery.com/jquery-1.11.0.min.js"></script>
-    <script src="http://netdna.bootstrapcdn.com/bootstrap/3.1.1/js/bootstrap.min.js"></script>
+	  <h2>Reverse your string!</h2>
+	  <form action ="{{ url_for('rev_str') }}" method="post">
+	    <input type="text" name="user_input">
+        <input type="submit" value="Reverse!">
+	  </form>
+	</div>
+   
   </body>
 </html>
